==> scripts/03_Analysis/01a_accumulation_curve.R <==
# Accumulation curves on genus, family, order on eDNA data


# Lib 
library(tidyverse)
library(ggplot2)
library(ggpubr)
library(conflicted)

# data
load("Rdata/02-clean-data.Rdata")

# 
'%ni%' <- Negate("%in%")
conflict_prefer("select", "dplyr")
conflict_prefer("mutate", "dplyr")
conflict_prefer("filter", "dplyr")

# Functions
source('scripts/03_Analysis/00_functions.R')


# First, remove the assignations above family (we can remove it in the 02_read script if necessary) --> keep it, put the families only in the supl. mat. 
list_read_step4 <- lapply(list_read_step4, function(x){
  x %>%
    #filter(new_rank_ncbi != "higher") %>%
    filter(province %in% c("Western_Indian_Ocean", "Southeast_Polynesia", "Tropical_Northwestern_Atlantic", "Western_Coral_Triangle", "Tropical_Southwestern_Pacific"))%>%
    filter(station %ni% c("estuaire_rio_don_diego_1", "estuaire_rio_don_diego_2", "estuaire_rio_don_diego_3", "glorieuse_distance_300m")) %>%
    filter(sample_method !="niskin" & comment %ni% c("Distance decay 600m", "Distance decay 300m"))%>%
    filter(project != "Curacao") %>%
    filter(habitat=="marine")%>%
    filter(habitat_type %ni% c("BAIE"))%>%
    filter(depth<40) %>%
    filter(family_name_corrected %ni% "Salmonidae")
})

# On the df as well
df_all_filters <- df_all_filters %>%
  filter(province %in% c("Western_Indian_Ocean", "Southeast_Polynesia", "Tropical_Northwestern_Atlantic", "Western_Coral_Triangle", "Tropical_Southwestern_Pacific"))%>%
  filter(station %ni% c("estuaire_rio_don_diego_1", "estuaire_rio_don_diego_2", "estuaire_rio_don_diego_3", "glorieuse_distance_300m")) %>%
  filter(sample_method !="niskin" & comment %ni% c("Distance decay 600m", "Distance decay 300m"))%>%
  filter(project != "Curacao") %>%
  filter(habitat=="marine")%>%
  filter(habitat_type %ni% c("BAIE"))%>%
  filter(site35!="") %>%
  filter(depth<40) %>%
  filter(family_name_corrected %ni% "Salmonidae")

# Re-format at region scale
# -----# After LULU 
df_all_filters_temp <- do.call("rbind", list_read_step4) 

# Split by region
list_read_step4 <- split(df_all_filters_temp, df_all_filters_temp$province)

# Counts
lapply(list_read_step4, function(x){
  length(unique(x$sequence))
})

lapply(list_read_step4, function(x){
  length(unique(x$station))
})

lapply(list_read_step4, function(x){
  length(unique(x$family_name_corrected))
})



# ------------------------------------------------------------------------------- # 
#### On family ----
# ------------------------------------------------------------------------------- # 

# rank_specify
rank_choice = 'family_name_corrected'

# accumlation all plots
liste_accumulation <- lapply(list_read_step4, accumulation_curve_df, species_unit = rank_choice)

# Asymptote of all plots 
liste_asymptote <- lapply(list_read_step4, asymptote_mm, species_unit = rank_choice)

# Unlist
df_accumulation <- bind_rows(liste_accumulation, .id = "project_name")
df_asymptote <- bind_rows(liste_asymptote, .id = "project_name")

# Dataset for plot
df_accumulation_all <- left_join(df_accumulation, df_asymptote, by = "project_name") %>%
  group_by(project_name) %>%
  mutate(position_asymptote_y = 0.20 * max(asymptote), 
         position_asymptote_x = max(sites))

# Plots
plot_acc_family <- ggplot(df_accumulation_all) + 
  geom_ribbon(aes(x = sites, ymin = richness-sd, ymax = richness+sd), fill = "lightblue", alpha = 0.8) +
  geom_line(aes(x = sites, y = richness)) +
  geom_hline(aes(yintercept = asymptote), linetype = "dashed", size = 1, col = "lightblue") +
  geom_text(aes(x = position_asymptote_x, y =position_asymptote_y, hjust = 1, label = paste("asymptote =", round(asymptote, 1), "Families")), col = "black") +
  facet_wrap(~project_name, scales = "free") +
  ylab("Number of MOTUs") +
  xlab("Samples (filter)") +
  theme_bw()

plot_acc_family

# Save
# ggsave("outputs/03_accumulation_curves/01b_accumulation_curve_all_projects_family.png", plot_acc_all, width = 12, height = 8)

# Add All samples
# Add a global saturation curve, i.e. all samples together?
all_accumulation <- accumulation_curve_df(df_all_filters, species_unit = rank_choice) %>%
  mutate(project_name = "All") %>%
  select(project_name, richness, sd, sites)


# Asymptote of all plots 
all_asymptote <- asymptote_mm(df_all_filters, species_unit = rank_choice) %>%
  mutate(project_name = "All") %>%
  select(project_name, asymptote, slope)


# Bind together
df_all_accumulation <- rbind(df_accumulation, all_accumulation)
df_all_asymptote <- rbind(df_asymptote, all_asymptote)

# 
df_join_all <- df_all_accumulation %>%
  left_join(., df_all_asymptote, by = "project_name") %>%
  group_by(project_name) %>%
  mutate(position_asymptote_y = 1.05*asymptote, 
         position_asymptote_x = max(sites),
         position_slope_y = 0.50 * max(asymptote)) 

# Plots
plot_acc_all <- ggplot(df_join_all, 
                       aes(fill = project_name, col = project_name)) + 
  geom_ribbon(aes(x = sites, ymin = richness-sd, ymax = richness+sd), alpha = 0.5) +
  geom_line(aes(x = sites, y = richness)) +
  geom_hline(aes(yintercept = asymptote, col = project_name), linetype = "dashed", size = 0.5) +
  ylab("Number of Families") +
  xlab("Samples (filter)") +
  theme_bw() # + scale_x_continuous(trans='log2') 

plot_acc_all

# ggsave("outputs/03_accumulation_curves/01_accumulation_curve_all_projects_combination_no_facet_family.png", plot_acc_all, width = 12, height = 8)

# Plot with facet
colnames(df_join_all) <- c("Province", "richness", "sd", "sites", "asymptote", "slope", "position_asymptote_y", "position_asymptote_x", "position_slope_y")
plot_acc_family <- ggplot(df_join_all, aes(fill = Province)) + 
  geom_ribbon(aes(x = sites, ymin = richness-sd, ymax = richness+sd),  alpha = 0.7) +
  geom_line(aes(x = sites, y = richness)) +
  geom_hline(aes(yintercept = asymptote), linetype = "dashed", size = 1) +
  scale_fill_manual(values=c("grey", "#a6611a", "#E5A729","#b2182b", "#80cdc1", "#015462"))+ 
  facet_wrap(~Province, scales = "free") +
  ylab("Number of families") +
  xlab("Samples (filter)") +
  theme_bw() + 
  ggtitle("Families") + 
  geom_text(aes(x = position_asymptote_x, y =position_asymptote_y, hjust = 1, label = paste(round(asymptote, 1), "Families")), col = "black", size=3)+
  geom_text(aes(x = position_asymptote_x, y =position_slope_y, hjust = 1, label = paste("Lomolino slope=",round(slope, 1))), col = "black", size=3)

plot_acc_family

ggsave("outputs/03_accumulation_curves/01b_accumulation_curve_all_projects_combination_no_facet_family.png", plot_acc_family, width = 12, height = 8)

# Table
# 

stats_family <- df_join_all %>%
  group_by(Province) %>%
  summarise(richness = max(richness), 
            asymptote = round(max(asymptote), 0)) %>%
  mutate(rank = 'Family')

# Save
write.csv(stats_family, "outputs/03_accumulation_curves/asymptotes_family.csv", row.names = F)

# Simple plot on all 
df_fam <- df_join_all %>% 
  filter(Province == "All") 

save(df_fam, file = "Rdata/accumulation_asymptote_families_all.rdata")

family <- ggplot(df_fam, aes(fill = Province, col = Province)) + 
  geom_ribbon(aes(x = sites, ymin = richness-sd, ymax = richness+sd),  alpha = 0.5) +
  geom_line(aes(x = sites, y = richness)) +
  geom_hline(aes(yintercept = asymptote), linetype = "dashed", size = 1) +
  ylab("Number of families") +
  xlab("Samples (filter)") +
  theme_bw() + 
  ggtitle("Families") + 
  geom_text(aes(x = position_asymptote_x, y =position_asymptote_y, hjust = 1, label = paste("asymptote =", round(asymptote, 0), "Families")), col = "black") 

family



# ------------------------------------------------------------------------------- # 
#### On MOTUs ----
# ------------------------------------------------------------------------------- # 

# rank_specify
rank_choice = 'sequence'

# accumlation all plots
liste_accumulation <- lapply(list_read_step4, accumulation_curve_df, species_unit = rank_choice)

# Asymptote of all plots 
liste_asymptote <- lapply(list_read_step4, asymptote_mm, species_unit = rank_choice)

# Unlist
df_accumulation <- bind_rows(liste_accumulation, .id = "project_name")
df_asymptote <- bind_rows(liste_asymptote, .id = "project_name")

# Dataset for plot
df_accumulation_all <- left_join(df_accumulation, df_asymptote, by = "project_name") %>%
  group_by(project_name) %>%
  mutate(position_asymptote_y = 0.20 * max(asymptote), 
         position_asymptote_x = max(sites))

# Add All samples
# Add a global saturation curve, i.e. all samples together?
all_accumulation_motu <- accumulation_curve_df(df_all_filters, species_unit = rank_choice) %>%
  mutate(project_name = "All") %>%
  select(project_name, richness, sd, sites)


# Asymptote of all plots 
all_asymptote_motu <- asymptote_mm(df_all_filters, species_unit = rank_choice) %>%
  mutate(project_name = "All") %>%
  select(project_name, asymptote, slope)


# Bind together
df_all_accumulation <- rbind(df_accumulation, all_accumulation_motu)
df_all_asymptote <- rbind(df_asymptote, all_asymptote_motu)

# 
df_join_all <- df_all_accumulation %>%
  left_join(., df_all_asymptote, by = "project_name") %>%
  group_by(project_name) %>%
  mutate(position_asymptote_y = 1.05*asymptote, 
         position_asymptote_x = max(sites),
         position_slope_y = 0.30 * max(asymptote)) 

# Plot with facet
colnames(df_join_all) <- c("Province", "richness", "sd", "sites", "asymptote", "slope", "position_asymptote_y", "position_asymptote_x", "position_slope_y")
plot_acc_motus <- ggplot(df_join_all, aes(fill = Province)) + 
  geom_ribbon(aes(x = sites, ymin = richness-sd, ymax = richness+sd),  alpha = 0.7) +
  geom_line(aes(x = sites, y = richness)) +
  geom_hline(aes(yintercept = asymptote), linetype = "dashed", size = 1) +
  facet_wrap(~Province, scales = "free") +
  scale_fill_manual(values=c("grey", "#a6611a", "#E5A729","#b2182b", "#80cdc1", "#015462"))+ 
  ylab("Number of MOTUs") +
  xlab("Samples (filter)") +
  theme_bw() + 
  ggtitle("MOTUs") + 
  geom_text(aes(x = position_asymptote_x, y =position_asymptote_y, hjust = 1, label = paste(round(asymptote, 1), "MOTUs")), col = "black", size=3)+
  geom_text(aes(x = position_asymptote_x, y =position_slope_y, hjust = 1, label = paste("Lomolino slope=",round(slope, 1))), col = "black", size=3)

plot_acc_motus

ggsave("outputs/03_accumulation_curves/01b_accumulation_curve_all_projects_combination_no_facet_motus.png", width = 12, height = 8)

# Table

stats<- df_join_all %>%
  group_by(Province) %>%
  summarise(richness = max(richness), 
            asymptote = round(max(asymptote), 0)) %>%
  mutate(rank = 'MOTUs')

# Save
write.csv(stats, "outputs/03_accumulation_curves/asymptotes_motus.csv", row.names = F)

# Simple plot on all 
df_motus <- df_join_all %>% 
  filter(Province == "All") 

save(df_motus, file = "Rdata/accumulation_asymptote_motus_all.rdata")

motus <- ggplot(df_motus, aes(fill = Province, col = Province)) + 
  geom_ribbon(aes(x = sites, ymin = richness-sd, ymax = richness+sd),  alpha = 0.5) +
  geom_line(aes(x = sites, y = richness)) +
  geom_hline(aes(yintercept = asymptote), linetype = "dashed", size = 1) +
  ylab("Number of MOTUs") +
  xlab("Samples (filter)") +
  theme_bw() + 
  ggtitle("MOTUs") + 
  geom_text(aes(x = position_asymptote_x, y =position_asymptote_y, hjust = 1, label = paste("asymptote =", round(asymptote, 0), "MOTUs")), col = "black") 

motus

# --------------------------------------------------------------------- # 
#### Final figure - combine all levels  ----
# --------------------------------------------------------------------- # 


ggarrange(plot_acc_motus, plot_acc_family, nrow=2, labels=c("a", "b"), common.legend = T, legend = "right")
ggsave("outputs/00_Figures_for_paper/Extended_Data/ED_figure5.png", width=10, height = 12)




# ------------------------------------------------------------------------------- # 
#### On genus ----
# ------------------------------------------------------------------------------- # 

# rank_specify
rank_choice = 'genus_name_corrected'

# accumlation all plots
liste_accumulation <- lapply(list_read_step4, accumulation_curve_df, species_unit = rank_choice)

# Asymptote of all plots 
liste_asymptote <- lapply(list_read_step4, asymptote_mm, species_unit = rank_choice)

# Unlist
df_accumulation <- bind_rows(liste_accumulation, .id = "project_name")
df_asymptote <- bind_rows(liste_asymptote, .id = "project_name")

# Dataset for plot
df_accumulation_all <- left_join(df_accumulation, df_asymptote, by = "project_name") %>%
  group_by(project_name) %>%
  mutate(position_asymptote_y = 0.20 * max(asymptote), 
         position_asymptote_x = max(sites))

# Add All samples
# Add a global saturation curve, i.e. all samples together?
all_accumulation <- accumulation_curve_df(df_all_filters, species_unit = rank_choice) %>%
  mutate(project_name = "All") %>%
  select(project_name, richness, sd, sites)

# Asymptote of all plots 
all_asymptote <- asymptote_mm(df_all_filters, species_unit = rank_choice) %>%
  mutate(project_name = "All") %>%
  select(project_name, asymptote)

# Bind together
df_all_accumulation <- rbind(df_accumulation, all_accumulation)
df_all_asymptote <- rbind(df_asymptote, all_asymptote)

# 
df_join_all <- df_all_accumulation %>%
  left_join(., df_all_asymptote, by = "project_name") %>%
  group_by(project_name) %>%
  mutate(position_asymptote_y = 0.20 * max(asymptote), 
         position_asymptote_x = max(sites)) 

# Plot with facet
plot_acc_genus <- ggplot(df_join_all, aes(fill = project_name)) + 
  geom_ribbon(aes(x = sites, ymin = richness-sd, ymax = richness+sd),  alpha = 0.7) +
  geom_line(aes(x = sites, y = richness)) +
  geom_hline(aes(yintercept = asymptote), linetype = "dashed", size = 1) +
  scale_fill_manual(values=c("grey", "#E5A729", "#80cdc1", "#a6611a", "#b2182b", "#015462"))+ 
  facet_wrap(~project_name, scales = "free") +
  ylab("Number of genus") +
  xlab("Samples (filter)") +
  theme_bw() + 
  ggtitle("Genus") + 
  geom_text(aes(x = position_asymptote_x, y =position_asymptote_y, hjust = 1, label = paste("asymptote =", round(asymptote, 1), "Genus")), col = "black") 

plot_acc_genus

ggsave("outputs/03_accumulation_curves/01b_accumulation_curve_all_projects_combination_no_facet_genus.png", width = 12, height = 8)

# Table
stats<- df_join_all %>%
  group_by(Province) %>%
  summarise(richness = max(richness), 
            asymptote = round(max(asymptote), 0)) %>%
  mutate(rank = 'Genus')

# Save
write.csv(stats, "outputs/03_accumulation_curves/asymptotes_genus.csv", row.names = F)

# Simple plot on all 
df_genus <- df_join_all %>% 
  filter(project_name == "All") %>%
  mutate(level = "genus")

genus <- ggplot(df_genus, aes(fill = Province, col = Province)) + 
  geom_ribbon(aes(x = sites, ymin = richness-sd, ymax = richness+sd),  alpha = 0.5) +
  geom_line(aes(x = sites, y = richness)) +
  geom_hline(aes(yintercept = asymptote), linetype = "dashed", size = 1) +
  ylab("Number of genus") +
  xlab("Samples (filter)") +
  theme_bw() + 
  ggtitle("Genus") + 
  geom_text(aes(x = position_asymptote_x, y =position_asymptote_y, hjust = 1, label = paste("asymptote =", round(asymptote, 0), "Genus")), col = "black") 

genus
