==> scripts/03_Analysis/04b_family_proportion_coverage_resolution.R <==
library(tidyverse)
library(dplyr)
library(reshape2)
library(vegan)
library(ggplot2)
library(cowplot)
library(grid)
library(gridExtra)
library(ggpubr)
library(RColorBrewer)



load("Rdata/02-clean-data.Rdata")
'%ni%' <- Negate("%in%")
df_all_filters <- df_all_filters %>%
  filter(province %in% c("Western_Indian_Ocean", "Southeast_Polynesia", "Tropical_Northwestern_Atlantic", "Western_Coral_Triangle", "Tropical_Southwestern_Pacific"))%>%
  filter(station %ni% c("estuaire_rio_don_diego_1", "estuaire_rio_don_diego_2", "estuaire_rio_don_diego_3", "glorieuse_distance_300m")) %>%
  filter(sample_method !="niskin" & comment %ni% c("Distance decay 600m", "Distance decay 300m"))%>%
  filter(project != "Curacao") %>%
  filter(habitat=="marine")%>%
  filter(habitat_type %ni% c("BAIE"))%>%
  filter(site35!="")%>%
  filter(depth<40) %>%
  filter(family_name_corrected %ni% "Salmonidae")



# load data to plot together

load("Rdata/nb_motus_per_family_global.Rdata")
load("Rdata/family_proportion_global.Rdata")
#load("Rdata/proportion_motus_similarity_threshold.Rdata")
family_coverage <- read.csv("outputs/01_read_data_stats/family_resolution_coefs.csv")
family <- unique(df_all_filters$family_name_corrected)
family_coverage <- subset(family_coverage, Family%in%family)


# select families with > 5 motus for main and < 5 motus for ED

count_families_global <- arrange(count_families_global, n_motus)
count_families_global_main <- count_families_global%>%
  subset(n_motus >= 5)
count_families_global_ED <- count_families_global %>%
  subset(n_motus < 5)
families_main <- as.character(unique(count_families_global_main$family))
families_ED <- as.character(unique(count_families_global_ED$family))

# Format to order properly

family_coverage_main <- family_coverage %>%
  subset(Family%in%families_main)
family_coverage_ED <- family_coverage %>%
  subset(Family%in%families_ED)

colnames(families_prop_global) <- c("family", "Southeast_Polynesia", "Tropical_Northwestern_Atlantic", "Western_Indian_Ocean", "Tropical_SouthWestern_Pacific", "Western_Coral_Triangle")
families_prop_global_main <- families_prop_global %>%
  subset(family%in%families_main)
families_prop_global_main <- reshape2::melt(families_prop_global_main)
colnames(families_prop_global_main) <- c("family", "Province", "prop")
save(families_prop_global_main, file="Rdata/family_proportion_region_main.rdata")

families_prop_global_ED <- families_prop_global %>%
  subset(family%in%family_coverage_ED$Family)
families_prop_global_ED <- reshape2::melt(families_prop_global_ED)
colnames(families_prop_global_ED) <- c("family", "Province", "prop")


order_main <- as.data.frame(levels(reorder(families_prop_global_main$family, families_prop_global_main$prop)))
colnames(order_main) <- "fam"
order_main <- order_main %>%
  subset(fam %in% family_coverage_main$Family)
order_ED <- as.data.frame(levels(reorder(families_prop_global_ED$family, families_prop_global_ED$prop)))
colnames(order_ED) <- "fam"
order_ED <- order_ED %>%
  subset(fam %in% family_coverage_ED$Family)

family_coverage_main$order<- order(order_main$fam)
family_coverage_ED$order<- order(order_ED$fam)



# plot each panel of the first figure
prop <- ggplot(families_prop_global_main, aes(x=reorder(family, prop), y = prop, fill = Province)) + 
  geom_bar(stat="identity", show.legend = TRUE) + 
  theme_bw() +
  scale_fill_manual(values =c("#a6611a", "#E5A729",  "#015462", "#b2182b", "#80cdc1"))+ 
  labs(title="Proportion of MOTUs at global scale, \nand their distribution in provinces", x="", y="")+ 
  theme(legend.position = "none")+
  theme(plot.title = element_text(size = 8, face="bold"), plot.margin=unit(c(0.1,0.2,0.6,0), "cm"))+
  coord_flip()

coverage <- ggplot(family_coverage_main, aes(x=order, y = coef_sequencing)) + 
  geom_bar(stat="identity", show.legend = TRUE) + 
  theme_bw() +
  labs(title="Proportion of sequences \nknown in databases", x="", y="")+ 
  theme(legend.position = "none")+
  theme(plot.title = element_text(size = 8, face = "bold"), plot.margin=unit(c(0.1,0.1,0.6,0), "cm"))+
  theme(axis.text.y = element_blank())+
  scale_x_continuous(breaks=family_coverage_main$order, labels=family_coverage_main$Family, expand = c(0,0))+
  ylim(0, 1)+
  coord_flip()

resolution <- ggplot(family_coverage_main, aes(x=order, y = coef_resolution)) + 
  geom_bar(stat="identity", show.legend = TRUE) + 
  theme_bw() +
  labs(title="Proportion of resolutive \nsequences in databases", x="", y="")+ 
  theme(legend.position = "none")+
  theme(plot.title = element_text(size = 8, face = "bold"), plot.margin=unit(c(0.1,0.1,0.6,0), "cm"))+
  theme(axis.text.y = element_blank())+
  scale_x_continuous(breaks=family_coverage_main$order, labels=family_coverage_main$Family, expand = c(0,0))+
  scale_y_continuous(breaks = c(0, 0.5, 1))+
  coord_flip()



plot_all_main <- ggarrange(prop, coverage, resolution, ncol=3, nrow=1, widths = c(1.5,1,1), labels = c("a", "b", "c"), common.legend = TRUE, legend="bottom")
plot_all_main
ggsave("outputs/00_Figures_for_paper/Extended_Data/ED_Figure2.png", width = 10.1, height = 8)


## plot each panel for second figure
prop <- ggplot(families_prop_global_ED, aes(x=reorder(family, prop), y = prop, fill = Province)) + 
  geom_bar(stat="identity", show.legend = TRUE) + 
  theme_bw() +
  scale_fill_manual(values =c("#a6611a", "#E5A729",  "#015462", "#b2182b", "#80cdc1"))+ 
  labs(title="Proportion of MOTUs at global scale, \nand their distribution in provinces", x="", y="")+ 
  theme(legend.position = "none")+
  ylim(0, 0.003)+
  theme(plot.title = element_text(size = 8, face="bold"), plot.margin=unit(c(0.1,0.2,0.6,0), "cm"))+
  coord_flip()

coverage <- ggplot(family_coverage_ED, aes(x=order, y = coef_sequencing)) + 
  geom_bar(stat="identity", show.legend = TRUE) + 
  theme_bw() +
  labs(title="Proportion of sequences \nknown in databases", x="", y="")+ 
  theme(legend.position = "none")+
  theme(plot.title = element_text(size = 8, face = "bold"), plot.margin=unit(c(0.1,0.1,0.6,0), "cm"))+
  theme(axis.text.y = element_blank())+
  scale_x_continuous(breaks=family_coverage_ED$order, labels=family_coverage_ED$Family, expand = c(0,0))+
  scale_y_continuous(breaks = c(0, 0.5, 1))+
  coord_flip()

resolution <- ggplot(family_coverage_ED, aes(x=order, y = coef_resolution)) + 
  geom_bar(stat="identity", show.legend = TRUE) + 
  theme_bw() +
  labs(title="Proportion of resolutive \nsequences in databases", x="", y="")+ 
  theme(legend.position = "none")+
  theme(plot.title = element_text(size = 8, face = "bold"), plot.margin=unit(c(0.1,0.1,0.6,0), "cm"))+
  theme(axis.text.y = element_blank())+
  scale_x_continuous(breaks=family_coverage_ED$order, labels=family_coverage_ED$Family, expand = c(0,0))+
  scale_y_continuous(breaks = c(0, 0.5, 1))+
  coord_flip()



plot_all_ED <- ggarrange(prop, coverage, resolution, ncol=3, nrow=1, widths = c(1.5,1,1), labels = c("a", "b", "c"), common.legend = TRUE, legend="bottom")
plot_all_ED
ggsave("outputs/00_Figures_for_paper/Extended_Data/ED_Figure3.png", width = 10.1, height = 10)
